refactoring products page and modularizing it
