+++ src/app/pages/product-detail/product-detail.component.html
