+++ src/app/pages/products/components/product/product.component.scss
