+++ src/app/pages/products/components/products/product.model.ts
