+++ src/app/pages/products/products-routing.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule, Router } from '@angular/router';
+import { ProductsComponent } from './components/products/products.component';
+
+const routes: Routes = [
+  { path: '', component: ProductsComponent },
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forChild(routes),
+  ],
+  exports: [
+    RouterModule,
+  ],
+})
+
+export class ProductsRoutingModule {}
