+++ src/app/pages/products/products.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { ProductComponent } from './components/product/product.component';
+import { ProductsComponent } from './components/products/products.component';
+
+import { ProductsRoutingModule } from './products-routing.module';
+
+@NgModule({
+  declarations: [
+    ProductComponent,
+    ProductsComponent,
+  ],
+  imports: [
+    CommonModule,
+    ProductsRoutingModule,
+  ],
+})
+
+export class ProductsModule {}
